ini Praktikum03

+++ Praktikum03/ContohTipeData15.java
@@ -9,10 +9,10 @@ public class ContohTipeData15 {
         long saldo = 150000000;
         int angkaDesimal = 0x10;
 
-        System.out.println("Golongan darah\t: " + (byte)golonganDarah);
+        System.out.println("Golongan darah\t: " + golonganDarah);
         System.out.println("Jumlah Penduduk\t: " + jumlahPenduduk);
         System.out.println("Suhu\t\t: " + suhu);
-        System.out.println("Berat\t\t: " + (float)berat);
+        System.out.println("Berat\t\t: " + berat);
         System.out.println("Saldo\t\t: " + saldo);
         System.out.println("Angka desimal\t: " + angkaDesimal);
     }
